+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - made `<pre>` text selectable / copyable
 - scrolling hack (view all content)
 - handle multiple file extensions to run (on editor save)
+- added `trinity:run` command
 
 0.3.0 / 2015-07-29
 ------------------